Move cleanup_experiment
