+++ packages/engine/src/experiment/controller/run.rs
@@ -1,5 +1,6 @@
 use std::{pin::Pin, sync::Arc, time::Duration};
 
+use memory::shared_memory::MemoryId;
 use tracing::Instrument;
 
 use crate::{
@@ -17,10 +18,12 @@ use crate::{
         package::ExperimentPackage,
     },
     output::{
-        local::LocalOutputPersistence, none::NoOutputPersistence, OutputPersistenceCreatorRepr,
+        buffer::RELATIVE_PARTS_FOLDER, local::LocalOutputPersistence, none::NoOutputPersistence,
+        OutputPersistenceCreatorRepr,
     },
-    proto::{EngineStatus, ExperimentRunTrait, PackageConfig},
+    proto::{EngineStatus, ExperimentId, ExperimentRunTrait, PackageConfig},
     simulation::package::creator::PackageCreators,
+    worker::runner::python::cleanup_python_runner,
     workerpool,
     workerpool::{comms::terminate::TerminateSend, WorkerPoolController},
     Error as CrateError,
@@ -345,4 +348,39 @@ fn worker_pool_exit_logic(
     };
 
     false
+}
+
+#[derive(PartialEq, Eq, Clone)]
+pub enum EngineExitStatus {
+    Success,
+    Error,
+}
+
+/// Shared memory cleanup in the process hard crash case.
+/// Not required for pod instances.
+pub fn cleanup_experiment(experiment_id: &ExperimentId, exit_status: EngineExitStatus) {
+    MemoryId::clean_up(experiment_id);
+
+    cleanup_python_runner(experiment_id, exit_status);
+
+    remove_experiment_parts(experiment_id);
+}
+
+// We might want to move the "/parts" folder to a temporary folder (like "/tmp" or "/var/tmp").
+fn remove_experiment_parts(experiment_id: &ExperimentId) {
+    let path = format!("{RELATIVE_PARTS_FOLDER}/{experiment_id}");
+    match std::fs::remove_dir_all(&path) {
+        Ok(_) => {
+            tracing::trace!(
+                experiment = %experiment_id,
+                "Removed parts folder for experiment {experiment_id}: {path:?}"
+            );
+        }
+        Err(err) => {
+            tracing::warn!(
+                experiment = %experiment_id,
+                "Could not clean up {path:?}: {err}"
+            );
+        }
+    }
 }
